requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
boto3==1.26.124
botocore==1.29.124
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2.1
django-cors-headers==3.14.0
django-storages==1.13.2
idna==3.4
jmespath==1.0.1
paramiko==3.1.0
psycopg2==2.9.6
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.30.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
uWSGI==2.0.21